+++ testData/org/elixir_lang/psi/scope/call_definition_clause/variants/callback.ex
@@ -1,0 +1,8 @@
+defmodule Behaviour do
+  @callback function_callback() :: term()
+  @macrocallback macro_callback(term()) :: term()
+
+  def helper do
+    call<caret>
+  end
+end